[fix] Add script to take over image cache

+++ scripts/evacuation-cache.js
@@ -6,7 +6,7 @@ const cacheDir = __dirname.includes('node_modules\\genshin-manager')
   : path.join(__dirname, '..', 'genshin-manager-cache')
 const assetCacheFolderPath = path.resolve(__dirname, '..', 'cache')
 
-if (!fs.existsSync(cacheDir)) {
+if (!fs.existsSync(cacheDir) && fs.existsSync(assetCacheFolderPath)) {
   fs.mkdirSync(cacheDir)
   fs.moveSync(
     path.join(assetCacheFolderPath, 'Images'),
